[refactor] Support Spark 3.1.x
