--- src/main/scala/tech/odes/utils/Logging.scala
@@ -1,7 +1,0 @@
-package tech.odes.utils
-
-import org.slf4j.LoggerFactory
-
-trait Logging {
-  final val log = LoggerFactory.getLogger(this.getClass.getName.stripSuffix("$"))
-}
